+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.15.1] - 2026-09-24
+
+- [#136](https://github.com/itk-dev/devops_itksites/pull/136)
+  Fall back to Drupal module versions for `drupal/*` packages when a detection
+  result has no Composer packages
+
 ## [1.15.0] - 2026-09-23
 
 - [#134](https://github.com/itk-dev/devops_itksites/pull/134)
@@ -371,7 +377,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-09-15
 
-[Unreleased]: https://github.com/itk-dev/devops_itksites/compare/1.15.0...HEAD
+[Unreleased]: https://github.com/itk-dev/devops_itksites/compare/1.15.1...HEAD
+[1.15.1]: https://github.com/itk-dev/devops_itksites/compare/1.15.0...1.15.1
 [1.15.0]: https://github.com/itk-dev/devops_itksites/compare/1.14.0...1.15.0
 [1.14.0]: https://github.com/itk-dev/devops_itksites/compare/1.13.0...1.14.0
 [1.13.0]: https://github.com/itk-dev/devops_itksites/compare/1.12.1...1.13.0